+++ tests/integration/wireless/v1/test_rate_plan.py
@@ -85,6 +85,7 @@ class RatePlanTestCase(IntegrationTestCase):
                         "sid": "WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "usage_notification_method": "POST",
                         "usage_notification_url": "https://callback.com",
+                        "data_limit_strategy": "block",
                         "url": "https://wireless.twilio.com/v1/RatePlans/WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ]
@@ -133,6 +134,7 @@ class RatePlanTestCase(IntegrationTestCase):
                 "sid": "WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "usage_notification_method": "POST",
                 "usage_notification_url": "https://callback.com",
+                "data_limit_strategy": "block",
                 "url": "https://wireless.twilio.com/v1/RatePlans/WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''
@@ -162,6 +164,7 @@ class RatePlanTestCase(IntegrationTestCase):
                 "unique_name": "unique_name",
                 "data_enabled": true,
                 "data_limit": 1000,
+                "data_limit_strategy": "block",
                 "data_metering": "payg",
                 "date_created": "2019-07-30T20:00:00Z",
                 "date_updated": "2019-07-30T20:00:00Z",
@@ -225,6 +228,7 @@ class RatePlanTestCase(IntegrationTestCase):
                 "sid": "WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "usage_notification_method": "POST",
                 "usage_notification_url": "https://callback.com",
+                "data_limit_strategy": "block",
                 "url": "https://wireless.twilio.com/v1/RatePlans/WPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''