+++ twilio/rest/verify/v2/service/__init__.py
@@ -19,8 +19,7 @@ from twilio.rest.verify.v2.service.verification_check import VerificationCheckLi
 
 
 class ServiceList(ListResource):
-    """ PLEASE NOTE that this class contains beta products that are subject to
-    change. Use them with caution. """
+    """  """
 
     def __init__(self, version):
         """
@@ -179,8 +178,7 @@ class ServiceList(ListResource):
 
 
 class ServicePage(Page):
-    """ PLEASE NOTE that this class contains beta products that are subject to
-    change. Use them with caution. """
+    """  """
 
     def __init__(self, version, response, solution):
         """
@@ -219,8 +217,7 @@ class ServicePage(Page):
 
 
 class ServiceContext(InstanceContext):
-    """ PLEASE NOTE that this class contains beta products that are subject to
-    change. Use them with caution. """
+    """  """
 
     def __init__(self, version, sid):
         """
@@ -359,8 +356,7 @@ class ServiceContext(InstanceContext):
 
 
 class ServiceInstance(InstanceResource):
-    """ PLEASE NOTE that this class contains beta products that are subject to
-    change. Use them with caution. """
+    """  """
 
     def __init__(self, version, payload, sid=None):
         """